requirements.txt
anthropic>=0.49.0
pyzotero>=1.5.0
pymupdf>=1.24.0
pymupdf4llm>=0.0.17
requests>=2.31.0
python-dotenv>=1.0.0
